Drawer component with react state
