+++ packages/nextjs/components/scaffold-eth/index.tsx
@@ -4,3 +4,5 @@ export { default as Balance } from "./Balance";
 export { default as AddressInput } from "./AddressInput";
 export { default as toast } from "./toast";
 export { default as ContractUI } from "./Contract/ContractUI";
+export { default as NavLinks } from "./Navlinks";
+export { default as Drawer } from "./Drawer";
